requirements.txt
groq>=0.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0

